Update Fibonacci to handle long values

+++ week-05/day-4/src/fibonacci/Fibonacci.java
@@ -2,7 +2,7 @@ package fibonacci;
 
 public class Fibonacci {
 
-  public int fibonacci(int index) {
+  public long fibonacci(int index) {
     if (index < 0) {
       return -1;
     }
